# services/torrent-engine/internal/domain/file_ref.go
package domain

type FileRef struct {
	Index          int    `json:"index"`
	Path           string `json:"path"`
	Length         int64  `json:"length"`
	BytesCompleted int64  `json:"bytesCompleted"`
	PieceStart     int    `json:"pieceStart,omitempty"` // inclusive
	PieceEnd       int    `json:"pieceEnd,omitempty"`   // exclusive
}